Update Module Issues in Mentorship portal (#5388)
* Update mentorship portal issues


* Update code


---------
