+++ frontend/__tests__/a11y/components/ModuleIssues.a11y.test.tsx
@@ -2,20 +2,19 @@ import { useQuery } from '@apollo/client/react'
 import { render } from '@testing-library/react'
 import { axe } from 'jest-axe'
 import { useTheme } from 'next-themes'
-import IssuesPage from 'app/my/mentorship/programs/[programKey]/modules/[moduleKey]/issues/page'
+import ModuleIssues from 'components/cards/ModuleIssues'
 
 jest.mock('@apollo/client/react', () => ({
   useQuery: jest.fn(),
 }))
 
 jest.mock('next/navigation', () => ({
-  useParams: jest.fn(() => ({ programKey: 'prog1', moduleKey: 'mod1' })),
-  useRouter: jest.fn(),
+  useRouter: jest.fn(() => ({ push: jest.fn(), replace: jest.fn() })),
   useSearchParams: jest.fn(() => new URLSearchParams()),
 }))
 
 const mockModuleData = {
-  getModule: {
+  managementModule: {
     name: 'Test Module',
     issues: [
       {
@@ -34,8 +33,19 @@ const mockModuleData = {
           },
         ],
       },
+      {
+        id: '2',
+        objectID: '2',
+        number: 102,
+        title: 'Second Issue Title',
+        state: 'open',
+        isMerged: false,
+        labels: ['docs'],
+        assignees: [],
+        taskDeadline: '2099-01-30T00:00:00Z',
+      },
     ],
-    issuesCount: 1,
+    issuesCount: 2,
     availableLabels: ['bug', 'feature-request', 'documentation'],
   },
 }
@@ -43,7 +53,7 @@ const mockModuleData = {
 describe.each([
   { theme: 'light', name: 'light' },
   { theme: 'dark', name: 'dark' },
-])('IssuesPage Accessibility ($name theme)', ({ theme }) => {
+])('ModuleIssues Accessibility ($name theme)', ({ theme }) => {
   beforeEach(() => {
     ;(useTheme as jest.Mock).mockReturnValue({ theme, setTheme: jest.fn() })
     document.documentElement.classList.toggle('dark', theme === 'dark')
@@ -59,7 +69,7 @@ describe.each([
       error: null,
     })
 
-    const { container } = render(<IssuesPage />)
+    const { container } = render(<ModuleIssues programKey="prog1" moduleKey="mod1" />)
 
     const results = await axe(container)
     expect(results).toHaveNoViolations()
