+++ frontend/src/app/mentorship/programs/[programKey]/page.tsx
@@ -71,6 +71,10 @@ const ProgramDetailsPage = () => {
     },
   ]
 
+  const summaryDetailsGridClass = program.description
+    ? 'grid grid-cols-1 gap-x-6 md:grid-cols-3'
+    : 'grid grid-cols-1 gap-x-6'
+
   return (
     <BreadcrumbStyleProvider className="bg-white dark:bg-[#212529]">
       <PageWrapper>
@@ -82,9 +86,15 @@ const ProgramDetailsPage = () => {
           showProgramActions={false}
         />
 
-        <Summary summary={program.description} />
+        <div className={summaryDetailsGridClass}>
+          <Summary summary={program.description} className="md:col-span-2" />
 
-        <Metadata details={programDetails} detailsTitle="Program Details" />
+          <Metadata
+            details={programDetails}
+            detailsTitle="Program Details"
+            className="md:col-span-1"
+          />
+        </div>
 
         <Tags tags={program.tags ?? undefined} domains={program.domains ?? undefined} />
